confirm password

+++ hamrogarden/src/components/login/login.js
@@ -8,6 +8,7 @@ const Login = () => {
   // const dataa = localStorage.getItem("userData");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [confirm, setConfirm] = useState("");
 
   const login = (e) => {
     e.preventDefault();
@@ -15,33 +16,37 @@ const Login = () => {
       email,
       password,
     };
-    axios
-      .post("http://localhost:7000/user/login", loginData)
-      .then((result) => {
-        if (result.status === 201) {
-          const token = result.data.token;
-          localStorage.setItem("token", token);
+    if (password === confirm) {
+      axios
+        .post("http://localhost:7000/user/login", loginData)
+        .then((result) => {
+          if (result.status === 201) {
+            const token = result.data.token;
+            localStorage.setItem("token", token);
 
-          window.alert("Login Success");
+            window.alert("Login Success");
 
-          // user
-          const userData = result.data.user;
-          localStorage.setItem("userData", userData);
-          // console.log(userData);
-          // console.log(
-          //   "User:",
-          //   result.data.user.name,
-          //   result.data.user.email,
-          //   result.data.user.password
-          // );
-          window.location.replace("/");
-        }
-      })
-      .catch((error) => {
-        console.log(error);
-        const err = error.response.data.message;
-        window.alert(err);
-      });
+            // user
+            const userData = result.data.user;
+            localStorage.setItem("userData", userData);
+            // console.log(userData);
+            // console.log(
+            //   "User:",
+            //   result.data.user.name,
+            //   result.data.user.email,
+            //   result.data.user.password
+            // );
+            window.location.replace("/");
+          }
+        })
+        .catch((error) => {
+          console.log(error);
+          const err = error.response.data.message;
+          window.alert(err);
+        });
+    } else {
+      window.alert("Password and Confirm password do not match!");
+    }
   };
 
   return (
@@ -68,6 +73,13 @@ const Login = () => {
                   placeholder="Password"
                   value={password}
                   onChange={(e) => setPassword(e.target.value)}
+                />
+                <input
+                  type="password"
+                  name=""
+                  placeholder="Confirm password"
+                  value={confirm}
+                  onChange={(e) => setConfirm(e.target.value)}
                 />
 
                 <button className="loginbtn" type="submit" onClick={login}>
